+++ CHANGELOG.md
@@ -29,7 +29,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
--
+
+## [7.2.0]
+
+### Added
+
+- Add NLPO Legrand contactor
+- Add NLD Legrand Double On/Off dimmer remote
+- Add NLFE Legrand On-Off dimmer switch
+- Add BTicino device support for BNCX, BNDL, BNSL
 
 ## [7.1.1]